requirements.txt
# Core dependencies
numpy>=1.20.0
pandas>=1.3.0
scikit-learn>=1.0.0
psutil>=5.9.0
requests>=2.28.0
watchdog>=2.1.0
joblib>=1.1.0
python-dotenv>=1.0.0  # For .env file support

# Optional dependencies
matplotlib>=3.5.0  # For visualization
flask>=2.0.0       # For web interface (if implemented)
pyyaml>=6.0        # For configuration files

# Raspberry Pi specific
RPi.GPIO>=0.7.0    # For GPIO control on Raspberry Pi